feat: modularize `ci_testing` and convert it into a plugin
